Install logrotate
Currently the rsyslog container does not log rotation. This caused
one implementation hitting over 30gb in logs that need to be manually
cleaned (and a full disk).

Rsyslog installed here already comes with a logrotate configuration
that is placed in the /etc/logrorate.d directory with a default weekly
clean up schedule. The only problem is that logrotate itself is not
installed, so no log rotation is occuring.